feat(context): add `getRootRef` action
Adds the `getRootRef` that deprecates the `rootRef` property of the context "action" object. This was done because the actions object should, conventionally, only contain functions, and the `rootRef` object is not a function.
Also refactors version numbers in docstrings, from X.Y.Z to X.Y, since nothing should be added / removed / deprecated in a patch version anyways.